chore: refactored code into new files, check-sat-result and add-variable-constraint
