+++ src/utilities/handle-check-proof-validity/z3-logic/arith-law-logic/z3-arith-funcs.ts
@@ -8,6 +8,8 @@
 
 // @ts-expect-error z3-solver is not recognising 'sat' as a valid export
 import {init, sat} from 'z3-solver';
+import {checkSatResult} from './check-sat-result';
+import {addVariableConstraint} from './add-variable-constraint';
 import {
 	type ImpliesPartExpr,
 	type SplitOperatorType,
@@ -22,327 +24,12 @@ async function initialiseContext() {
 	return [Int, And, Solver, Not, Implies];
 }
 
-function checkSatResult({
-	result,
-	solver,
-	beforeImpliesExpr,
-	afterImpliesExpr,
-}: {
-	result: any;
-	solver: any;
-	beforeImpliesExpr: ImpliesPartExpr[];
-	afterImpliesExpr: ImpliesPartExpr[];
-}): boolean {
-	console.log('sat result overall is:', result);
-	if (result === 'sat') {
-		const model = solver.model();
-		const declarations = model.decls();
-
-		const modelValues: {
-			xValue: undefined | string;
-			yValue: undefined | string;
-			zValue: undefined | string;
-		} = {
-			xValue: undefined,
-			yValue: undefined,
-			zValue: undefined,
-		};
-		// Logging values satisfying z3 stack to console
-
-		modelValues.xValue =
-			declarations[0].name() === 'x'
-				? model.get(declarations[0]).asString()
-				: undefined;
-		console.log('model xValue:', modelValues.xValue);
-
-		if (declarations.length === 2) {
-			modelValues.yValue =
-				declarations[1].name() === 'y'
-					? model.get(declarations[1]).asString()
-					: undefined;
-			console.log('model yValue:', modelValues.yValue);
-		}
-
-		if (declarations.length === 3) {
-			modelValues.zValue =
-				declarations[2].name() === 'z'
-					? model.get(declarations[2]).asString()
-					: undefined;
-			console.log('model zValue:', modelValues.zValue);
-		}
-
-		let errorPresent = false;
-
-		let i = 0;
-
-		while (i < beforeImpliesExpr.length) {
-			const lhsValue = beforeImpliesExpr[i].value;
-			const lhsOperator = beforeImpliesExpr[i].operator;
-
-			const rhsValue = afterImpliesExpr[i].value;
-			const rhsOperator = afterImpliesExpr[i].operator;
-
-			// Map through all operators and values
-
-			if (lhsOperator === rhsOperator) {
-				let comparisonOperator: string;
-
-				switch (lhsOperator) {
-					case '=': {
-						comparisonOperator = '=';
-						break;
-					}
-
-					case '>': {
-						comparisonOperator = '>';
-						break;
-					}
-
-					case '<': {
-						comparisonOperator = '<';
-						break;
-					}
-
-					case '>=': {
-						comparisonOperator = '>=';
-						break;
-					}
-
-					case '<=': {
-						comparisonOperator = '<=';
-						break;
-					}
-				}
-
-				// Check xValue found to satisfy constraints is as expected for operator type, i.e. xValue === lhsValue === rhsValue when operator is '=', xValue > lhsValue and xValue > rhsValue when operator is '>' etc.
-
-				const compareOperations = {
-					'=': (x: number, y: number) => x === y,
-					'>': (x: number, y: number) => x > y,
-					'<': (x: number, y: number) => x < y,
-					'>=': (x: number, y: number) => x >= y,
-					'<=': (x: number, y: number) => x <= y,
-				};
-
-				let modelValue = '';
-
-				switch (i) {
-					case 0: {
-						if (modelValues.xValue) {
-							modelValue = modelValues.xValue;
-						}
-
-						break;
-					}
-
-					case 1: {
-						if (modelValues.yValue) {
-							modelValue = modelValues.yValue;
-						}
-
-						break;
-					}
-
-					case 2: {
-						if (modelValues.zValue) {
-							modelValue = modelValues.zValue;
-						}
-
-						break;
-					}
-
-					default: {
-						console.log(
-							`No applicable modelValue to set based on number i of ${i}`,
-						);
-						errorPresent = true;
-					}
-				}
-
-				if (modelValue === '') {
-					console.log("modelValue was not assigned a value, i.e. ''");
-					errorPresent = true;
-				}
-
-				if (
-					compareOperations[comparisonOperator as SplitOperatorType](
-						Number(modelValue),
-						Number(lhsValue),
-					) &&
-					compareOperations[comparisonOperator as SplitOperatorType](
-						Number(modelValue),
-						Number(rhsValue),
-					)
-				) {
-					console.log('pass');
-				} else {
-					console.log(
-						`model value ${modelValue} returned didnt meet logical constraint for variable`,
-					);
-					errorPresent = true;
-				}
-			} else {
-				console.log(
-					"Error: Values discovered to satisfy constraints for arith law call in skip law call proof don't match ones passed to method, hence not a display of validity in this instance",
-				);
-				errorPresent = true;
-			}
-
-			i++;
-		}
-
-		if (errorPresent) {
-			console.log(
-				'checkSatResult error: model values generated dont match constraints in proof, e.g. xValue = 5 with constraint of x < 3',
-			);
-			return false;
-		}
-
-		return true;
-	}
-
-	return false;
-}
-
-async function addVariableConstraint(
-	variableName: 'x' | 'y' | 'z',
-	operator: SplitOperatorType,
-	value: number,
-	Int: any,
-	And: any,
-	Not: any,
-	solver: any,
-): Promise<any | false> {
-	async function addVariableConstraintX(
-		value: number,
-		operator: SplitOperatorType,
-		And: any,
-		Not: any,
-	): Promise<any> {
-		const x = Int.const('x');
-		const constraint = getConstraint(x, value, operator, And, Not);
-		return constraint;
-	}
-
-	async function addVariableConstraintY(
-		value: number,
-		operator: SplitOperatorType,
-		And: any,
-		Not: any,
-	): Promise<any> {
-		const y = Int.const('y');
-		const constraint = getConstraint(y, value, operator, And, Not);
-		return constraint;
-	}
-
-	async function addVariableConstraintZ(
-		value: number,
-		operator: SplitOperatorType,
-		And: any,
-		Not: any,
-	): Promise<any> {
-		const z = Int.const('z');
-		const constraint = getConstraint(z, value, operator, And, Not);
-		return constraint;
-	}
-
-	let addConstraint;
-
-	// Handle Z3 constant variable to create
-	switch (variableName) {
-		case 'x': {
-			addConstraint = await addVariableConstraintX(value, operator, And, Not);
-			break;
-		}
-
-		case 'y': {
-			addConstraint = await addVariableConstraintY(value, operator, And, Not);
-			break;
-		}
-
-		case 'z': {
-			addConstraint = await addVariableConstraintZ(value, operator, And, Not);
-			break;
-		}
-	}
-
-	if (addConstraint === false || addConstraint === undefined) {
-		console.error(
-			'Error constructing logical statements for z3 solver in addVariableConstraint',
-		);
-		return false;
-	}
-
-	solver.add(addConstraint);
-
-	const result = await solver.check();
-
-	if (result === 'sat') {
-		console.log(`Constraint added for ${variableName}: ${value}`);
-		return addConstraint;
-	}
-
-	console.error(`Error adding constraint for ${variableName}: ${value}`);
-	return false;
-	// Could do with sending message about needing to refresh the page to re-try adding all assertions to z3 stack again due to error encountered
-}
-
-async function getConstraint(
-	passedVarName: any,
-	value: number,
-	operator: SplitOperatorType,
-	And: any,
-	Not: any,
-): Promise<{firstStatement: any}> {
-	let firstStatement;
-
-	switch (operator) {
-		case '=': {
-			firstStatement = And(
-				passedVarName.gt(value - 1),
-				passedVarName.lt(value + 1),
-			).eq(true); // Should be the exact same bounds as in firstStatement
-			break;
-		}
-
-		case '<': {
-			firstStatement = passedVarName.lt(value).eq(true);
-			break;
-		}
-
-		case '>': {
-			firstStatement = passedVarName.gt(value).eq(true);
-			break;
-		}
-
-		case '<=': {
-			firstStatement = Not(passedVarName.gt(value)).eq(true);
-			break;
-		}
-
-		case '>=': {
-			firstStatement = Not(passedVarName.lt(value)).eq(true);
-			break;
-		}
-	}
-
-	return firstStatement;
-}
-
 async function constructImpliesExpr(
 	And: any,
 	constraints: any[],
 	solver: any,
 ): Promise<false | any> {
 	const impliesExprResult = And(...constraints).eq(true);
-
-	/* If (constraints.length === 2) {
-		impliesExprResult = And(constraints[0], constraints[1]).eq(true);
-	} else if (constraints.length === 3) {
-		impliesExprResult = And(constraints[0], constraints[1], constraints[2]).eq(
-			true,
-		);
-	} */
 
 	solver.add(impliesExprResult);
